# Conflicts:
#	src/main/java/seng302/Controllers/RaceViewController.java
#	src/main/java/seng302/Controllers/StarterController.java
#	src/main/java/seng302/Controllers/TableController.java
#	src/main/java/seng302/Model/Boat.java
#	src/main/java/seng302/Model/Competitor.java
#	src/main/java/seng302/Model/DataReceiver.java
#	src/main/java/seng302/Model/MutablePoint.java
#	src/main/java/seng302/Model/RaceEvent.java
#	src/main/java/seng302/Parsers/RaceStatusParser.java
#	src/main/java/seng302/Parsers/RaceXMLParser.java
#	src/main/java/seng302/TestMockDatafeed/DataSender.java
#	src/main/resources/raceView.fxml
#	src/main/resources/startView.fxml
#	src/main/resources/table.fxml

+++ src/main/java/seng302/Parsers/RaceStatusData.java
@@ -8,10 +8,10 @@ import java.util.HashMap;
  */
 public class RaceStatusData {
 
-    private Integer currentTime;
+    private long currentTime;
     private Integer raceID;
     private String raceStatus;
-    private Integer expectedStartTime;
+    private long expectedStartTime;
     private Double windDirection;
     private Integer windSpeed;
     private Integer numBoatsInRace;
@@ -19,7 +19,7 @@ public class RaceStatusData {
     private HashMap<Integer, BoatStatus> boatStatuses = new HashMap<>();
 
 
-    public RaceStatusData(Integer currentTime, Integer raceID, String raceStatus, Integer expectedStartTime, Double windDirection, Integer windSpeed, Integer numBoatsInRace, Integer raceType, HashMap<Integer, BoatStatus> boatStatuses) {
+    public RaceStatusData(long currentTime, Integer raceID, String raceStatus, long expectedStartTime, Double windDirection, Integer windSpeed, Integer numBoatsInRace, Integer raceType, HashMap<Integer, BoatStatus> boatStatuses) {
         this.currentTime = currentTime;
         this.raceID = raceID;
         this.raceStatus = raceStatus;
@@ -32,7 +32,7 @@ public class RaceStatusData {
     }
 
 
-    public Integer getCurrentTime() {
+    public long getCurrentTime() {
         return currentTime;
     }
 
@@ -44,7 +44,7 @@ public class RaceStatusData {
         return raceStatus;
     }
 
-    public Integer getExpectedStartTime() {
+    public long getExpectedStartTime() {
         return expectedStartTime;
     }
